# Conflicts:
#	package.json
#	packages/size-check/package.json
#	packages/uni-api/package.json
#	packages/uni-app-harmony/package.json
#	packages/uni-app-plus/package.json
#	packages/uni-app-uts/package.json
#	packages/uni-app-vite/package.json
#	packages/uni-app-vue/package.json
#	packages/uni-app/package.json
#	packages/uni-automator/package.json
#	packages/uni-cli-shared/package.json
#	packages/uni-cli-utils/package.json
#	packages/uni-cloud/package.json
#	packages/uni-components/package.json
#	packages/uni-core/package.json
#	packages/uni-h5-vite/package.json
#	packages/uni-h5-vue/package.json
#	packages/uni-h5/package.json
#	packages/uni-i18n/package.json
#	packages/uni-mp-alipay/package.json
#	packages/uni-mp-baidu/package.json
#	packages/uni-mp-compiler/package.json
#	packages/uni-mp-core/package.json
#	packages/uni-mp-jd/package.json
#	packages/uni-mp-kuaishou/package.json
#	packages/uni-mp-lark/package.json
#	packages/uni-mp-qq/package.json
#	packages/uni-mp-toutiao/package.json
#	packages/uni-mp-vite/package.json
#	packages/uni-mp-vue/package.json
#	packages/uni-mp-weixin/package.json
#	packages/uni-mp-xhs/package.json
#	packages/uni-nvue-styler/package.json
#	packages/uni-preprocess/package.json
#	packages/uni-push/package.json
#	packages/uni-quickapp-webview/package.json
#	packages/uni-runtime/package.json
#	packages/uni-shared/package.json
#	packages/uni-stacktracey/package.json
#	packages/uni-stat/dist/uni-cloud-stat.cjs.js
#	packages/uni-stat/dist/uni-cloud-stat.es.js
#	packages/uni-stat/dist/uni-stat.cjs.js
#	packages/uni-stat/dist/uni-stat.es.js
#	packages/uni-stat/package.json
#	packages/uni-stat/src/utils/index.js
#	packages/uni-stat/src/utils/util.js
#	packages/uni-uts-v1/package.json
#	packages/uni-vue-devtools/package.json
#	packages/uni-vue/package.json
#	packages/uts-darwin-arm64/package.json
#	packages/uts-darwin-x64/package.json
#	packages/uts-linux-x64-gnu/package.json
#	packages/uts-linux-x64-musl/package.json
#	packages/uts-win32-ia32-msvc/package.json
#	packages/uts-win32-x64-msvc/package.json
#	packages/uts/package.json
#	packages/vite-plugin-uni/package.json
#	pnpm-lock.yaml

+++ scripts/release.js
@@ -125,19 +125,25 @@ async function main() {
 
 function updateVersions(version) {
   // 1. update root package.json
-  updatePackage(path.resolve(__dirname, '..'), version)
+  updatePackage(path.resolve(__dirname, '..'), version, true)
   // 2. update all packages
   packages.forEach((p) => updatePackage(getPkgRoot(p), version))
 }
 
-function updatePackage(pkgRoot, version) {
+function updatePackage(pkgRoot, version, ignoreDeps = false) {
   const pkgPath = path.resolve(pkgRoot, 'package.json')
+  if (!fs.existsSync(pkgPath)) {
+    console.log(colors.yellow(`Cannot find package.json in ${pkgRoot}`))
+    return
+  }
   const pkg = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'))
   pkg.version = version
-  updateDeps(pkg, 'dependencies', version)
-  updateDeps(pkg, 'devDependencies', version)
-  updateDeps(pkg, 'peerDependencies', version)
-  updateDeps(pkg, 'optionalDependencies', version)
+  if (!ignoreDeps) {
+    updateDeps(pkg, 'dependencies', version)
+    updateDeps(pkg, 'devDependencies', version)
+    updateDeps(pkg, 'peerDependencies', version)
+    updateDeps(pkg, 'optionalDependencies', version)
+  }
   fs.writeFileSync(pkgPath, JSON.stringify(pkg, null, 2) + '\n')
 }
 
@@ -163,6 +169,10 @@ async function publishPackage(pkgName, version, runIfNotDry) {
   }
   const pkgRoot = getPkgRoot(pkgName)
   const pkgPath = path.resolve(pkgRoot, 'package.json')
+  if (!fs.existsSync(pkgPath)) {
+    console.log(colors.yellow(`Cannot find package.json in ${pkgRoot}`))
+    return
+  }
   const pkg = JSON.parse(fs.readFileSync(pkgPath, 'utf-8'))
   if (pkg.private) {
     return
